fix: widen the wind overlay's speed-bucket counter past 255
The perf rewrite counted each bucket's points in a Uint8List, and a whole
6400-particle population can land in one bucket under strong wind — the
count then wraps at 255, dropping the bucket (or most of it) so new
particles vanish and stale trails outlive a rotation. Count in 16 bits,
and make the streak tests actually see the particles: the sampled
boundary was Scaffold's white one (blank overlays passed), and the z7
viewport held too few particles to trip the wrap. A zoomed-in Taiwan
field now puts thousands of points in one bucket, pinning the count at
300+ bright pixels — the buggy build measures ~60.

+++ lib/features/map/presentation/widgets/wind_particle_overlay.dart
@@ -241,7 +241,12 @@ class _TrailBuffer {
   );
 
   /// Live point count per bucket this frame.
-  final Uint8List _counts = Uint8List(16);
+  ///
+  /// 16 bits, not 8: the whole population (6400) can land in one speed bucket
+  /// under strong wind, and an 8-bit counter wraps at 255 — the bucket then
+  /// draws the wrong point count (or none at all, when the count wraps to 0),
+  /// which reads as particles vanishing and stale trails outliving a rotation.
+  final Uint16List _counts = Uint16List(16);
 
   /// Floats per bucket: the largest population (6400 at z3) as x,y pairs.
   static const int _bucketCapacity = 6400 * 2;
